VIITE-765 after revert is called check if road part was automatically reserved and release it if yes
Split project service spec to multiple files (project stuff, links stuff, tr stuff)
