+++ digiroad2-viite/src/test/scala/fi/liikennevirasto/viite/ProjectServiceTRSpec.scala
@@ -1,0 +1,130 @@
+package fi.liikennevirasto.viite
+
+import java.net.ConnectException
+import java.util.Properties
+
+import fi.liikennevirasto.digiroad2.oracle.OracleDatabase
+import fi.liikennevirasto.digiroad2.{DigiroadEventBus, RoadLinkService}
+import fi.liikennevirasto.viite.dao.{Discontinuity, ProjectDAO, ProjectState, RoadAddressProject}
+import org.apache.http.client.config.RequestConfig
+import org.apache.http.client.methods.HttpGet
+import org.apache.http.conn.{ConnectTimeoutException, HttpHostConnectException}
+import org.apache.http.impl.client.HttpClientBuilder
+import org.joda.time.DateTime
+import org.mockito.Mockito.reset
+import org.scalatest.mock.MockitoSugar
+import org.scalatest.{BeforeAndAfter, FunSuite, Matchers}
+import slick.driver.JdbcDriver.backend.Database
+import slick.driver.JdbcDriver.backend.Database.dynamicSession
+import slick.jdbc.StaticQuery.interpolation
+
+class ProjectServiceTRSpec extends FunSuite with Matchers with BeforeAndAfter {
+  val properties: Properties = {
+    val props = new Properties()
+    props.load(getClass.getResourceAsStream("/digiroad2.properties"))
+    props
+  }
+  val mockProjectService= MockitoSugar.mock[ProjectService]
+  val mockRoadLinkService = MockitoSugar.mock[RoadLinkService]
+  val mockRoadAddressService = MockitoSugar.mock[RoadAddressService]
+  val mockEventBus = MockitoSugar.mock[DigiroadEventBus]
+  val roadAddressService = new RoadAddressService(mockRoadLinkService, mockEventBus) {
+    override def withDynSession[T](f: => T): T = f
+
+    override def withDynTransaction[T](f: => T): T = f
+  }
+  val projectService = new ProjectService(roadAddressService, mockRoadLinkService, mockEventBus) {
+    override def withDynSession[T](f: => T): T = f
+
+    override def withDynTransaction[T](f: => T): T = f
+  }
+
+  val projectServiceWithRoadAddressMock= new ProjectService(mockRoadAddressService, mockRoadLinkService, mockEventBus) {
+    override def withDynSession[T](f: => T): T = f
+
+    override def withDynTransaction[T](f: => T): T = f
+  }
+
+  after {
+    reset(mockRoadLinkService)
+  }
+  def withDynTransaction[T](f: => T): T = OracleDatabase.withDynTransaction(f)
+  def runWithRollback[T](f: => T): T = {
+    Database.forDataSource(OracleDatabase.ds).withDynTransaction {
+      val t = f
+      dynamicSession.rollback()
+      t
+    }
+  }
+
+  val dr2properties: Properties = {
+    val props = new Properties()
+    props.load(getClass.getResourceAsStream("/digiroad2.properties"))
+    props
+  }
+
+  private def testConnection: Boolean = {
+    val url = dr2properties.getProperty("digiroad2.tierekisteriViiteRestApiEndPoint")
+    val request = new HttpGet(url)
+    request.setConfig(RequestConfig.custom().setConnectTimeout(2500).build())
+    val client = HttpClientBuilder.create().build()
+    try {
+      val response = client.execute(request)
+      try {
+        response.getStatusLine.getStatusCode >= 200
+      } finally {
+        response.close()
+      }
+    } catch {
+      case e: HttpHostConnectException =>
+        false
+      case e: ConnectTimeoutException =>
+        false
+      case e: ConnectException =>
+        false
+    }
+  }
+
+  test("fetch project data and send it to TR") {
+    assume(testConnection)
+    runWithRollback {
+      val project = RoadAddressProject(1, ProjectState.Incomplete, "testiprojekti", "Test", DateTime.now(), "Test",
+        DateTime.now(), DateTime.now(), "info", List(
+          ReservedRoadPart(5: Long, 203: Long, 203: Long, 5: Double, 5: Long, Discontinuity.apply("jatkuva"), 8: Long,
+            None: Option[DateTime], None: Option[DateTime])), None)
+      ProjectDAO.createRoadAddressProject(project)
+      sqlu""" insert into road_address_changes(project_id,change_type,new_road_number,new_road_part_number,new_track_code,new_start_addr_m,new_end_addr_m,new_discontinuity,new_road_type,new_ely) Values(1,1,6,1,1,0,10.5,1,1,8) """.execute
+      //Assuming that there is data to show
+      val responses = projectService.getRoadAddressChangesAndSendToTR(Set(1))
+      responses.projectId should be(1)
+    }
+  }
+
+  test("update ProjectStatus when TR saved") {
+    val sent2TRState = ProjectState.apply(2) //notfinnished
+    val savedState = ProjectState.apply(5)
+    val projectId = 0
+    val addresses = List(ReservedRoadPart(5: Long, 203: Long, 203: Long, 5: Double, 5: Long, Discontinuity.apply("jatkuva"), 8: Long, None: Option[DateTime], None: Option[DateTime]))
+    val roadAddressProject = RoadAddressProject(projectId, ProjectState.apply(2), "TestProject", "TestUser", DateTime.now(), "TestUser", DateTime.parse("1901-01-01"), DateTime.now(), "Some additional info", List(), None)
+    runWithRollback {
+      val saved = projectService.createRoadLinkProject(roadAddressProject)
+      val stateaftercheck = projectService.updateProjectStatusIfNeeded(sent2TRState, savedState, "", saved.id)
+      stateaftercheck.description should be(ProjectState.Saved2TR.description)
+    }
+
+  }
+
+  test("Update to TRerror state") {
+    val sent2TRState = ProjectState.apply(2) //notfinnished
+    val savedState = ProjectState.apply(3)
+    val projectId = 0
+    val addresses = List(ReservedRoadPart(5: Long, 203: Long, 203: Long, 5: Double, 5: Long, Discontinuity.apply("jatkuva"), 8: Long, None: Option[DateTime], None: Option[DateTime]))
+    val roadAddressProject = RoadAddressProject(projectId, ProjectState.apply(2), "TestProject", "TestUser", DateTime.now(), "TestUser", DateTime.parse("1901-01-01"), DateTime.now(), "Some additional info", List(), None)
+    runWithRollback {
+      val saved = projectService.createRoadLinkProject(roadAddressProject)
+      val stateaftercheck = projectService.updateProjectStatusIfNeeded(sent2TRState, savedState, "failed", saved.id)
+      stateaftercheck.description should be(ProjectState.ErroredInTR.description)
+    }
+
+  }
+}
